Adding EmployeeDAO file and refactoring Application.java to the proper packag
